# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fellowship Entrance Exam and Seat Allocation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /doctors:
    get:
      operationId: listDoctors
      tags: [doctors]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Doctor" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginInput:
      type: object
      properties:
        email: { type: string }
        password: { type: string }
      required: [email, password]
...
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user: { $ref: "#/components/schemas/CurrentUser" }
      required: [token, user]
    CurrentUser:
      type: object
      properties:
        id: { type: integer }
        email: { type: string }
        fullName: { type: string }
        role: { type: string }
        candidateId: { type: integer, nullable: true }
        unitId: { type: integer, nullable: true }
        programId: { type: integer, nullable: true }
      required: [id, email, fullName, role]
...
    Doctor:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        fullName: { type: string }
        email: { type: string }
        unitId: { type: integer, nullable: true }
        unitName: { type: string, nullable: true }
        assignedCandidates: { type: integer }
      required: [id, userId, fullName, email, assignedCandidates]
...